+++ framework/src/main/kotlin/dev/alpas/routing/Controller.kt
@@ -75,31 +75,3 @@ open class Controller {
         call.cookie.forget(name, path, domain)
     }
 }
-
-class ControllerMiddleware(
-    val methods: List<String>? = null,
-    private val middleware: KClass<out Middleware<HttpCall>>,
-    private vararg val others: KClass<out Middleware<HttpCall>>
-) {
-    val all by lazy { listOf(middleware, *others) }
-
-    constructor(
-        method: String,
-        middleware: KClass<out Middleware<HttpCall>>,
-        vararg others: KClass<out Middleware<HttpCall>>
-    ) : this(
-        methods = listOf(method),
-        middleware = middleware,
-        others = *others
-    )
-
-    constructor(middleware: KClass<out Middleware<HttpCall>>, vararg others: KClass<out Middleware<HttpCall>>) : this(
-        methods = null,
-        middleware = middleware,
-        others = *others
-    )
-
-    fun matches(method: String): Boolean {
-        return this.methods == null || this.methods.contains(method)
-    }
-}
